backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
networkx
psycopg2-binary
